Delete unused function
